Refactor all culinarytech, init v1.0
