+++ backend/app/routes/spoonacular/wine.py
@@ -17,7 +17,7 @@ router = APIRouter(
 API_KEY = os.getenv("SPOONACULAR_API_KEY")
 BASE_URL = "https://api.spoonacular.com"
 
-# Endpoint Dish pairing for wine | User search a recipe that goes well with a wine
-@router.get("/wine")
-def wine():
-    return {"message": "Wine endpoint"}
+# # Endpoint Dish pairing for wine | User search a recipe that goes well with a wine
+# @router.get("/wine")
+# def wine():
+#     return {"message": "Wine endpoint"}
