+++ backend/requirements.txt
@@ -7,3 +7,4 @@ requests
 passlib[bcrypt]
 python-multipart
 python-jose[cryptography]
+pydantic[email]
