+++ frontend/src/App.css
@@ -1,10 +1,10 @@
 @import "tailwindcss";
 
-/* @theme {
+@theme {
     --color-background: #FAFAFA;
     --color-primary: #FFFACD; 
     --color-accent: #C3DAE6;
     --color-secondary: #D4ECDD;
     --color-muted: #F8EDEB;
     --color-text: #1F1F1F;
-} */
+}
