+++ frontend/src/layout/static/Footer.jsx
