--- frontend/src/pages/NotFound.jsx
@@ -1,1 +1,0 @@
-// Error 404
