--- frontend/src/pages/Shop.jsx
@@ -1,14 +1,0 @@
-import Navbar from "../components/static/Navbar"
-import Footer from "../components/static/Footer"
-
-export default function Shop() {
-    return (
-        <div className="flex flex-col min-h-screen">
-            <Navbar />
-            <main className="flex-grow">
-
-            </main>
-            <Footer />
-        </div>
-    )
-}
