+++ frontend/src/pages/user/Forgot.jsx
@@ -1,6 +1,6 @@
 import ForgotForm from '../../components/auth/ForgotForm';
 import AuthCard from '../../components/ui/AuthCard';
-import React from 'react';
+
 
 const Forgot = () => {
     return (
