+++ frontend/src/pages/user/Login.jsx
@@ -1,7 +1,5 @@
 import LoginForm from '../../components/auth/LoginForm';
 import AuthCard from '../../components/ui/AuthCard';
-// import Navbar from '@/components/Navbar';
-// import Footer from '@/components/Footer';
 
 export default function Login() {
     return (
